+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Unreleased changes
 * (aanet/tohd5) new option to read run ID from header, not event. 
   in old versions of JTE, the event.run_id is overwritten with the default, 1.
 * there is now a new command line utility called ``streamds`` for non-pythonistas
+* The new ``km3pipe.ahrs`` now contains AHRS calibration routines
 
 Version 7
 ---------